Fixes issues that were found after full check.

+++ app/controllers/course_subscriptions_controller.rb
@@ -1,12 +1,14 @@
 class CourseSubscriptionsController < ApplicationController
-  before_action :authenticate_user!
+  before_filter :authenticate_user!
 
   def create
-    course.participants << current_user
+    return course.participants << current_user unless current_user.blocked_in?(course)
+    redirect_for_blocked_user("A blocked user can't subscribe on course.")
   end
 
   def destroy
-    course.course_users.where(user_id: current_user.id).first.destroy
+    return course.course_users.where(user_id: current_user.id).first.destroy unless current_user.blocked_in?(course)
+    redirect_for_blocked_user("A blocked user can't unsubscribe on course.")
   end
 
   private
@@ -14,5 +16,12 @@ class CourseSubscriptionsController < ApplicationController
   def course
     @course ||= Course.find(params[:course_id])
   end
+
+  def redirect_for_blocked_user(message)
+    flesh[:blocked_user_error] = message
+    flash.keep(:blocked_user_error)
+    render js: "window.location = #{courses_path.to_json}"
+  end
+
   helper_method :course
 end
